Improved and tested Batch Improved Json Creator Created Compare Pipelines

+++ src/main/java/org/bdcourse/stream/TwitterStreamWordCount.java
@@ -50,7 +50,7 @@ public class TwitterStreamWordCount {
 		DataStream<String> streamSource = null;
 		streamSource = env.addSource(twitterSource);
 
-		DataStream<Tuple2<String, Double>> tweets = streamSource
+		DataStream<Tuple2<String, Integer>> tweets = streamSource
 				.filter(new TweetContainingHashtag())
 				.flatMap(new SelectTweetsWithHashtags())
 				.filter(new FilterTweetsFromList())
